vinyl: do not release latch in the middle of vylog rotate
vy_log_rotate() releases the log latch between reading the last vylog
file and writing the new vylog file. This works as long as the latch
implementation guarantees that latch_lock() called immediately after
latch_unlock() on the same lock doesn't yield. Although this is true
now, we shouldn't rely on that, because this may change any time.