schema: improve arguments check in grant or revoke on universe
The name of the universe is optional, so we don't check it. If a user
wants to specify extra options in the grant, such as if_not_exists, and
mistakes object name argument with options argument, options are
silently ignored:

  box.schema.user.grant('tnt', 'read,write,execute', 'universe', {if_not_exists = true})

Fix this by adding Lua code that ensures that universe name is a scalar
(string or nil).

Closes #3146