vinyl: abort worker threads on shutdown
If the user terminates tarantool while compaction is in progress,
tarantool won't exit immediately - it will hang until all dump and
compaction tasks that are currently in progress are complete, which
may take quite a while on large data sets. What is especially funny,
once a task has been finished, vinyl will not commit the produced run
file in the vylog, because the event loop has already been stopped,
and so will delete it after restart and schedule the task anew.

This patch makes the scheduler forcefully abort all running tasks
as soon as possible.

Closes #3166