sql: remove unnecessary VDBE flags
This commit is a part of work which is necessary for
successful removal of OP_Transaction. During that
opcode execution it doesn't do any meaningful work,
however it presence was necessary, because without
that opcode VDBE flags like 'readOnly', 'bIsReader'
would be assigned incorrectly. Also unnecessary OP_Destroy
opcode was removed.

This flags are unnecessary for us, so in this patch
they were removed (and everything related to them).

For #3119