sql: ban of REINDEX syntax
This change removes REINDEX from keywords and also disables tests
which were using this syntax.

This patch affects only parser, REINDEX implementation still exists,
because it is planned to be fixed and enabled in the scope of #3195.

Note that REINDEX keyword is still in '%fallback' directive in
'parse.y'. Reserved keywords included there to avoid situation when
a keyword has no usages within 'parse.y' file.

Fixes #2174