sql: remove unused operands from OP_IdxInsert/Replace
And rework sqlite3CompleteInsertion. Before the patch it
* accepts actually unused pTab;
* array of records to insert in different indexes, but inserts
  only one;
* has sqlite code style.

Code style fix is obvious. PTab is used only to check, that it is
not a view and to check, that a first index is primary:
1) caller functions already guarantee the pTab is not a view;
2) regardless of a first index nature: primary or not - it is not
   used here. It is useless check. With the same success we can
   check this in each function, that uses struct Table.

Array of records to insert makes no sense, since insertion is
being done in a primary index only. It is enough to pass a
register with a primary index tuple.