sql: remove unnecessary MakeRecord opcodes
OP_MakeRecord in INSERT/UPDATE query is required for making
insertion into Tarantool and for checking constraints with
ON CONFLICT clause. There was one OP_MakeRecord for each
constraint with ON CONFLICT, and each constraint was checked by
VDBE. But now some constraints with error actions like ABORT can
be handled by Tarantool facilities (without emitting bytecode) and
therefore MakeRecord for that index would be unnecessary if it is
not a primary key index (in that case MakeRecord is also needed
for OP_IdxInsert). This patch contains a fix, which removes
unnecessary MakeRecord opcodes from INSERT/UPDATE VDBE listings.

Closes #3317