vinyl: force index rebuild if indexed field type is narrowed
If the type of an indexed field is narrowed (e.g. is_nullable flag is
cleared or the type is changed from 'integer' to 'unsigned'), but the
index structure remains the same (same fields, indexed in the same
order), we won't rebuild the index, instead we will check that all
tuples stored in the space conform to the new format (CheckSpaceFormat)
and if so modify the index in-place (ModifyIndex). This is OK for memtx,
but not for Vinyl, because even if the space contains no tuples
conflicting with the new format, there may be overwritten or deleted
statements, which can't be detected by CheckSpaceFormat.

That being said, let's make index_def_change_requires_rebuild() virtual
(add it to index_vtab) and force rebuild of Vinyl indexes if the type of
any indexed field is narrowed.