alter: introduce TruncateIndex AlterSpaceOp for truncating space
Space truncation is equivalent to dropping and recreating all its
indexes so we use DropIndex + CreateIndex for it. This works now, but
with the introduction of the ALTER preparation phase (see the next
commit), it won't. Anyway, abusing existing operations for truncation
looks rather ugly. Let's introduce a new operation for it.