sql: LIKE & GLOB pattern comparison issue
Currently function that compares pattern and string for GLOB & LIKE
operators doesn't work properly. It uses ICU reading function which
was assumed having other return codes and the implementation for the
comparison ending isn't paying attention to some special cases, hence
in those cases it works improperly.

With the patch applied an error will be returned in case there's an
invalid UTF-8 symbol in pattern & pattern containing only valid UTF-8
symbols will not be matched with the string that contains invalid
symbol.

Сloses #3251
Сloses #3334
Part of #3572