sql: increase time quota for selectG test on vinyl
The test case has execution time around 25 seconds on vinyl (while
around 5 seconds on memtx) on a slow laptop with HDD. On a modern laptop
with SSD it is ~8 seconds for vinyl and ~2-3 seconds for memtx. Both
measurement were performed with parallel run of many instances of the
test.

Also enabled ability to run a sql-tap test w/o test-run. It was broken
since b09f5964.

Fixes #3590.